+++ sklearn/neural_networks/tests/test_rbm.py
@@ -4,7 +4,7 @@ from numpy.testing import assert_almost_equal, assert_array_equal
 
 from sklearn.base import clone
 from sklearn.datasets import load_digits
-from sklearn.neural_networks import RestrictedBolzmannMachine
+from sklearn.neural_networks import BernoulliRBM
 
 np.seterr(all='warn')
 
@@ -16,8 +16,8 @@ Xdigits /= Xdigits.max()
 def test_fit():
     X = Xdigits.copy()
 
-    rbm = RestrictedBolzmannMachine(n_components=64, learning_rate=0.1,
-                                    n_particles=10, n_iter=7, random_state=9)
+    rbm = BernoulliRBM(n_components=64, learning_rate=0.1,
+                       n_particles=10, n_iter=7, random_state=9)
     rbm.fit(X)
 
     assert_almost_equal(rbm.pseudo_likelihood(X).mean(), -20., decimal=0)
@@ -29,8 +29,8 @@ def test_fit():
 def test_fit_transform():
     """Check proper implementation of fit_transform"""
     X = Xdigits[:100]
-    rbm1 = RestrictedBolzmannMachine(n_components=16, n_particles=5,
-                                     n_iter=5, random_state=42)
+    rbm1 = BernoulliRBM(n_components=16, n_particles=5,
+                        n_iter=5, random_state=42)
     rbm2 = clone(rbm1)
 
     Xt1 = rbm1.fit(X).transform(X)