+++ doc/whats_new.rst
@@ -386,7 +386,7 @@ Changelog
    - :class:`MiniBatchKMeans`: Add random reassignment of cluster centers
      with little observations attached to them, by `Gael Varoquaux`_.
 
-   - New transformer :ref:`neural_networks.RestrictedBoltzmannMachine`
+   - New transformer :ref:`neural_networks.BernoulliRBM`
      by `Yann Dauphin`_.